requirements.txt
Flask==3.0.3
requests
python-dotenv~=1.0.1
werkzeug==3.0.4
SQLAlchemy==2.0.35
schedule~=1.2.2
marshmallow~=3.23.2